☂ marry poppins: some unimportant change proposals

See merge request taz/taz-app!1084

+++ app/src/main/java/de/taz/app/android/data/DataService.kt
@@ -9,7 +9,6 @@ import de.taz.app.android.dataStore.StorageDataStore
 import de.taz.app.android.persistence.repository.*
 import de.taz.app.android.simpleDateFormat
 import de.taz.app.android.util.SingletonHolder
-import de.taz.app.android.util.runIfNotNull
 import kotlinx.coroutines.*
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
@@ -49,18 +48,15 @@ class DataService(applicationContext: Context) {
     }
 
     private suspend fun ensureIssueCount() = ensureCountLock.withLock {
-        runIfNotNull(
-            downloadIssueNumberLiveData.value,
-            maxStoredIssueNumberLiveData.value
-        ) { downloaded, max ->
+        val downloaded = downloadIssueNumberLiveData.value
+        val max = maxStoredIssueNumberLiveData.value
+        if (downloaded != null && max != null) {
             var downloadedCounter = downloaded
             while (downloadedCounter > max) {
-                runBlocking {
-                    issueRepository.getEarliestDownloadedIssueStub()?.let {
-                        contentService.deleteIssue(IssuePublication(it.issueKey))
-                    }
-                    downloadedCounter--
+                issueRepository.getEarliestDownloadedIssueStub()?.let {
+                    contentService.deleteIssue(IssuePublication(it.issueKey))
                 }
+                downloadedCounter--
             }
         }
     }
